Add files via upload
Map of the game area that shows where the player is as well as the walls. Loaded into the game (main.py opens map.txt) whereupon every line is added to a list. Main.txt then goes through this list and adds a wall or player depending on the symbol placement of the file